deploy/terraform/cloud_function/requirements.txt
functions-framework==3.*
google-cloud-bigquery==3.*
google-cloud-storage==2.*
google-cloud-logging==3.*
dbt-core==1.7.*
dbt-bigquery==1.7.*
